+++ .storybook/main.ts
@@ -1,5 +1,4 @@
 import type { StorybookConfig } from '@storybook/react-vite';
-
 const config: StorybookConfig = {
   stories: [
     '../src/**/*/stories.@(ts|tsx)',
@@ -15,6 +14,8 @@ const config: StorybookConfig = {
     name: '@storybook/react-vite',
     options: {},
   },
+  docs: {
+    autodocs: true,
+  },
 };
-
 export default config;
